Refactor Sanitizer polyfill
Turns out Chrome supports Sanitizer but has none of the sanitize
methods... It's only useful in `el.setHTML()`.

Refactor polyfill to utilize Chrome's implementations as best as
possible.
