+++ SanitizerBase.js
@@ -1,139 +1,7 @@
+import { nativeSupport, getSantizerUtils, sanitize, sanitizeFor, trustPolicies } from './sanitizerUtils.js';
 import { SanitizerConfig as defaultConfig } from './SanitizerConfigBase.js';
-import { nativeSupport, getSantizerUtils } from './sanitizerUtils.js';
-import { documentToFragment } from './dom.js';
-import { isObject, getType } from './utility.js';
-import { createPolicy } from './trust.js';
-import { urls } from './attributes.js';
 
 const protectedData = new WeakMap();
-const allowProtocols = ['https:'];
-
-if (! allowProtocols.includes(location.protocol)) {
-	allowProtocols.push(location.protocol);
-}
-
-export function sanitizeNode(node, config = defaultConfig) {
-	try {
-		if (! (node instanceof Node)) {
-			throw new TypeError(`Expected a Node but got a ${getType(node)}.`);
-		} else if (! isObject(config)) {
-			throw new TypeError(`Expected config to be an object but got ${getType(config)}.`);
-		}
-
-		const {
-			allowElements, allowComments, allowAttributes, allowCustomElements,
-			blockElements, dropAttributes, dropElements
-		} = config;
-
-		switch(node.nodeType) {
-			case Node.TEXT_NODE:
-				break;
-
-			case Node.ELEMENT_NODE: {
-				if (! (node.parentNode instanceof Node)) {
-					break;
-				}
-
-				const tag = node.tagName.toLowerCase();
-
-				if (Array.isArray(dropElements) && dropElements.includes(tag)) {
-					node.remove();
-				} else if (Array.isArray(blockElements) && blockElements.includes(tag)) {
-					if (node.hasChildNodes()) {
-						[...node.childNodes].forEach(node => sanitizeNode(node, config));
-						node.replaceWith(...node.childNodes);
-					} else {
-						node.remove();
-					}
-				} else if (tag.includes('-') && ! allowCustomElements) {
-					node.remove();
-				} else if (Array.isArray(allowElements) && ! allowElements.includes(tag)) {
-					node.remove();
-				} else if (tag === 'template') {
-					sanitizeNode(node.content, config);
-				} else {
-					if (node.hasAttributes()) {
-						node.getAttributeNames()
-							.forEach(attr => sanitizeNode(node.getAttributeNode(attr), config));
-					}
-
-					if (node.hasChildNodes()) {
-						[...node.childNodes].forEach(node => sanitizeNode(node, config));
-					}
-				}
-
-				break;
-			}
-
-			case Node.ATTRIBUTE_NODE: {
-				const { value, ownerElement } = node;
-				const name = node.name.toLowerCase();
-				const tag = ownerElement.tagName.toLowerCase();
-
-				if (
-					urls.includes(name)
-					&& ! allowProtocols.includes(new URL(value, document.baseURI).protocol)
-				) {
-					ownerElement.removeAttributeNode(node);
-				} else if (isObject(dropAttributes)) {
-					if (
-						name in dropAttributes
-						&& ['*', tag].some(sel => dropAttributes[name].includes(sel))
-					) {
-						ownerElement.removeAttributeNode(node);
-
-						if (name.startsWith('on')) {
-							delete ownerElement[name];
-						}
-					}
-				} else if (isObject(allowAttributes)) {
-					if (
-						! name.startsWith('data-')
-						&& ! (name in allowAttributes
-						&& ['*', tag].some(sel => allowAttributes[name].includes(sel)))
-					) {
-						ownerElement.removeAttributeNode(node);
-
-						if (name.startsWith('on')) {
-							delete ownerElement[name];
-						}
-					}
-				}
-
-				break;
-			}
-
-			case Node.COMMENT_NODE: {
-				if (! allowComments) {
-					node.remove();
-				}
-
-				break;
-			}
-
-			case Node.DOCUMENT_NODE:
-			case Node.DOCUMENT_FRAGMENT_NODE: {
-				if (node.hasChildNodes()) {
-					[...node.childNodes].forEach(node => sanitizeNode(node, config));
-				}
-
-				break;
-			}
-
-			case Node.CDATA_SECTION_NODE:
-			case Node.PROCESSING_INSTRUCTION_NODE:
-			case Node.DOCUMENT_TYPE_NODE:
-			default: {
-				node.parentElement.removeChild(node);
-			}
-		}
-	} catch(err) {
-		node.parentElement.removeChild(node);
-		console.error(err);
-	}
-
-	return node;
-}
 
 /**
  * Need to create a policy for the Sanitizer API since
@@ -141,7 +9,7 @@ export function sanitizeNode(node, config = defaultConfig) {
  * which would create infinite recursion.
  * @type {TrustedTypePolicy}
  */
-let rawPolicy = createPolicy('sanitizer-raw#html', { createHTML: input => input });
+
 
 /**
  * @SEE https://wicg.github.io/sanitizer-api/
@@ -169,23 +37,11 @@ export class Sanitizer {
 	}
 
 	sanitize(input) {
-		if (input instanceof Document) {
-			return this.sanitize(documentToFragment(input, { policy: rawPolicy }));
-		} else if (input instanceof DocumentFragment) {
-			/* It'd be great if this could be moved to a worker script... */
-			const frag = input.cloneNode(true);
-			sanitizeNode(frag, this.getConfiguration());
-
-			return frag;
-		}
+		return sanitize(input, { config: this.getConfiguration() });
 	}
 
 	sanitizeFor(tag, content) {
-		const el = document.createElement(tag);
-		const temp = document.createElement('template');
-		temp.innerHTML = rawPolicy.createHTML(content);
-		el.append(this.sanitize(temp.content));
-		return el;
+		return sanitizeFor(tag, content, { config: this.getConfiguration() });
 	}
 
 	static getDefaultConfiguration() {
@@ -194,7 +50,4 @@ export class Sanitizer {
 }
 
 const { setHTML, polyfill } = getSantizerUtils(Sanitizer, defaultConfig);
-
-export const createHTML = input => rawPolicy.createHTML(input);
-export const trustPolicies = [rawPolicy.name];
-export { nativeSupport, setHTML, polyfill, allowProtocols };
+export { nativeSupport, setHTML, polyfill, trustPolicies };
