+++ test/index.html
@@ -8,7 +8,7 @@
 		<meta name="description" content="A testing page for my JavaScript library" />
 		<meta name="robots" content="nofollow,noindex" />
 		<base href="/test/" />
-		<meta http-equiv="Content-Security-Policy-x" content="
+		<meta http-equiv="Content-Security-Policy" content="
 			default-src 'self';
 			base-uri 'self';
 			manifest-src 'self';
