+++ trust-policies.js
@@ -97,7 +97,17 @@ export const sanitizerConfig = {
 export const sanitizer = new Sanitizer(sanitizerConfig);
 
 export const getDefaultPolicy = callOnce(() => createPolicy('default', {
-	createHTML: input => sanitizer.sanitizeFor('div', input).innerHTML,
+	createHTML: input => {
+		if (sanitizer.sanitize instanceof Function && sanitizer.sanitizeFor instanceof Function) {
+			return sanitizer.sanitizeFor('div', input).innerHTML
+		} else if (Element.prototype.setHTML instanceof Function) {
+			const el = document.createElement('div');
+			el.setHTML(input, { sanitizer });
+			return el.innerHTML;
+		} else {
+			return input;
+		}
+	},
 	createScript: () => trustedTypes.emptyScript.toString(),
 	createScriptURL: input => {
 		const url = new URL(input, document.baseURI);
